Renamed tokenIds.js to settings.js
